+++ src/cryptonote_basic/src/transaction_prefix.rs
@@ -3,7 +3,7 @@ use crate::tx_out::TxOut;
 use crate::tx_in::TxIn;
 
 pub struct TransactionPrefix {
-    version: u8,
+    version: usize,
     unlock_time: u64,
     vin: Vec<TxIn>,
     vout: Vec<TxOut>,
